backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow>=10.2.0
python-decouple==3.8
# psycopg2-binary==2.9.9  # Optional, only needed for PostgreSQL
django-filter>=24.0
drf-yasg==1.21.7
setuptools>=65.0.0
pymongo>=4.6.0
django-mongodb-backend>=5.2.0
